+++ src/main/java/com/zyc/zdh/controller/ProjectController.java
@@ -166,6 +166,7 @@ public class ProjectController extends BaseController {
     public ReturnInfo<ProjectInfo> project_detail(String id) {
         try {
             ProjectInfo projectInfo = projectMapper.selectByPrimaryKey(id);
+            checkPermissionByProductAndDimGroup(zdhPermissionService, projectInfo.getProduct_code(), projectInfo.getDim_group());
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(), "查询成功", projectInfo);
         } catch (Exception e) {
             return ReturnInfo.build(RETURN_CODE.FAIL.getCode(), "查询失败", e);
@@ -187,6 +188,10 @@ public class ProjectController extends BaseController {
         try {
 
             ProjectInfo oldProjectInfo = projectMapper.selectByPrimaryKey(projectInfo.getId());
+
+            checkPermissionByProductAndDimGroup(zdhPermissionService, projectInfo.getProduct_code(), projectInfo.getDim_group());
+            checkPermissionByProductAndDimGroup(zdhPermissionService, oldProjectInfo.getProduct_code(), oldProjectInfo.getDim_group());
+
 
             JSONArray jsonArray=new JSONArray();
 
@@ -226,6 +231,7 @@ public class ProjectController extends BaseController {
     @Transactional(propagation= Propagation.NESTED)
     public ReturnInfo<ProjectInfo> project_add(ProjectInfo projectInfo,String[] enum_value, String[] enum_value_context) {
         try {
+            checkPermissionByProductAndDimGroup(zdhPermissionService, projectInfo.getProduct_code(), projectInfo.getDim_group());
 
             JSONArray jsonArray=new JSONArray();
 
@@ -262,6 +268,7 @@ public class ProjectController extends BaseController {
     @Transactional(propagation= Propagation.NESTED)
     public ReturnInfo project_delete(String[] ids) {
         try {
+            checkPermissionByProductAndDimGroup(zdhPermissionService, projectMapper, projectMapper.getTable(), ids);
             projectMapper.deleteLogicByIds("project_info",ids, new Timestamp(System.currentTimeMillis()));
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(), "删除成功", null);
         } catch (Exception e) {
