+++ src/main/java/com/zyc/zdh/controller/ZdhApprovalController.java
@@ -111,6 +111,7 @@ public class ZdhApprovalController extends BaseController{
     public ReturnInfo<ApprovalConfigInfo> approval_config_detail(String id) {
         try{
             ApprovalConfigInfo approvalConfigInfo=approvalConfigMapper.selectByPrimaryKey(id);
+            checkPermissionByProduct(zdhPermissionService, approvalConfigInfo.getProduct_code());
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(), "查询成功", approvalConfigInfo);
         }catch (Exception e){
              logger.error("类:"+Thread.currentThread().getStackTrace()[1].getClassName()+" 函数:"+Thread.currentThread().getStackTrace()[1].getMethodName()+ " 异常: {}", e);
@@ -161,8 +162,8 @@ public class ZdhApprovalController extends BaseController{
             checkPermissionByProduct(zdhPermissionService, approvalConfigInfo.getProduct_code());
             checkPermissionByProduct(zdhPermissionService, oldApprovalConfigInfo.getProduct_code());
 
-            BeanUtils.copyProperties(oldApprovalConfigInfo, approvalConfigInfo);
             approvalConfigInfo.setEmployee_id(getOwner());
+            approvalConfigInfo.setCreate_time(oldApprovalConfigInfo.getCreate_time());
             approvalConfigMapper.updateByPrimaryKeySelective(approvalConfigInfo);
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(), "更新成功", null);
         }catch (Exception e){
@@ -290,6 +291,10 @@ public class ZdhApprovalController extends BaseController{
     public ReturnInfo<ApprovalAuditorInfo> approval_auditor_update(ApprovalAuditorInfo approvalAuditorInfo) {
         try{
             ApprovalAuditorInfo old=approvalAuditorMapper.selectByPrimaryKey(approvalAuditorInfo.getId());
+
+            checkPermissionByProduct(zdhPermissionService, approvalAuditorInfo.getProduct_code());
+            checkPermissionByProduct(zdhPermissionService, old.getProduct_code());
+
             approvalAuditorInfo.setCreate_time(old.getCreate_time());
             approvalAuditorMapper.updateByPrimaryKeySelective(approvalAuditorInfo);
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(), "更新成功", approvalAuditorInfo);
@@ -310,6 +315,7 @@ public class ZdhApprovalController extends BaseController{
     public ReturnInfo<ApprovalAuditorInfo> approval_auditor_detail(String id) {
         try{
             ApprovalAuditorInfo approvalAuditorInfo=approvalAuditorMapper.selectById(id);
+            checkPermissionByProduct(zdhPermissionService, approvalAuditorInfo.getProduct_code());
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(), "查询成功", approvalAuditorInfo);
         }catch (Exception e){
              logger.error("类:"+Thread.currentThread().getStackTrace()[1].getClassName()+" 函数:"+Thread.currentThread().getStackTrace()[1].getMethodName()+ " 异常: {}", e);
@@ -367,6 +373,8 @@ public class ZdhApprovalController extends BaseController{
     @ResponseBody
     public ReturnInfo<ApprovalAuditorFlowInfo> approval_auditor_flow_add(ApprovalAuditorFlowInfo approvalAuditorFlowInfo) {
         try{
+            checkPermissionByProduct(zdhPermissionService, approvalAuditorFlowInfo.getProduct_code());
+
             List<ApprovalAuditorFlowInfo> approvalAuditorFlowInfos=approvalAuditorFlowMapper.select(approvalAuditorFlowInfo);
             if(approvalAuditorFlowInfos!=null && approvalAuditorFlowInfos.size()>0){
                 throw new Exception("审批节点以存在");
@@ -395,6 +403,7 @@ public class ZdhApprovalController extends BaseController{
     @Transactional(propagation= Propagation.NESTED)
     public ReturnInfo<Object> approval_auditor_flow_delete(String[] ids) {
         try{
+            checkPermissionByProductAndDimGroup(zdhPermissionService, approvalAuditorFlowMapper, approvalAuditorFlowMapper.getTable(), ids);
             approvalAuditorFlowMapper.deleteLogicByIds("approval_auditor_flow_info", ids, new Timestamp(System.currentTimeMillis()));
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(), "删除成功", null);
         }catch (Exception e){
@@ -415,6 +424,10 @@ public class ZdhApprovalController extends BaseController{
     public ReturnInfo<ApprovalAuditorFlowInfo> approval_auditor_flow_update(ApprovalAuditorFlowInfo approvalAuditorFlowInfo) {
         try{
             ApprovalAuditorFlowInfo old = approvalAuditorFlowMapper.selectByPrimaryKey(approvalAuditorFlowInfo.getId());
+
+            checkPermissionByProduct(zdhPermissionService, approvalAuditorFlowInfo.getProduct_code());
+            checkPermissionByProduct(zdhPermissionService, old.getProduct_code());
+
             approvalAuditorFlowInfo.setCreate_time(old.getCreate_time());
             approvalAuditorFlowInfo.setOwner(old.getOwner());
             approvalAuditorFlowInfo.setIs_delete(Const.NOT_DELETE);
@@ -439,6 +452,8 @@ public class ZdhApprovalController extends BaseController{
     public ReturnInfo<ApprovalAuditorFlowInfo> approval_auditor_flow_detail(String id) {
         try{
             ApprovalAuditorFlowInfo approvalAuditorFlowInfo=approvalAuditorFlowMapper.selectByPrimaryKey(id);
+            checkPermissionByProduct(zdhPermissionService, approvalAuditorFlowInfo.getProduct_code());
+
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(), "查询成功", approvalAuditorFlowInfo);
         }catch (Exception e){
             logger.error("类:"+Thread.currentThread().getStackTrace()[1].getClassName()+" 函数:"+Thread.currentThread().getStackTrace()[1].getMethodName()+ " 异常: {}", e);
@@ -478,6 +493,8 @@ public class ZdhApprovalController extends BaseController{
     @ResponseBody
     public ReturnInfo<Object> approval_event_add(ApprovalEventInfo approvalEventInfo) {
         try{
+            checkPermissionByProduct(zdhPermissionService, approvalEventInfo.getProduct_code());
+
             List<ApprovalEventInfo> approvalEventInfos=approvalEventMapper.select(approvalEventInfo);
             if(approvalEventInfos!=null && approvalEventInfos.size()>0){
                 throw new Exception("审批事件已存在");
@@ -503,6 +520,7 @@ public class ZdhApprovalController extends BaseController{
     public ReturnInfo<ApprovalEventInfo> approval_event_detail(String id) {
         try{
             ApprovalEventInfo approvalEventInfo=approvalEventMapper.selectById(id);
+            checkPermissionByProduct(zdhPermissionService, approvalEventInfo.getProduct_code());
             return ReturnInfo.build(RETURN_CODE.SUCCESS.getCode(), "查询成功", approvalEventInfo);
         }catch (Exception e){
              logger.error("类:"+Thread.currentThread().getStackTrace()[1].getClassName()+" 函数:"+Thread.currentThread().getStackTrace()[1].getMethodName()+ " 异常: {}", e);
@@ -544,6 +562,10 @@ public class ZdhApprovalController extends BaseController{
     public ReturnInfo<ApprovalEventInfo> approval_event_update(ApprovalEventInfo approvalEventInfo) {
         try{
             ApprovalEventInfo old = approvalEventMapper.selectByPrimaryKey(approvalEventInfo.getId());
+
+            checkPermissionByProduct(zdhPermissionService, approvalEventInfo.getProduct_code());
+            checkPermissionByProduct(zdhPermissionService, old.getProduct_code());
+
             approvalEventInfo.setCreate_time(old.getCreate_time());
             approvalEventInfo.setUpdate_time(new Timestamp(System.currentTimeMillis()));
             approvalEventMapper.updateByPrimaryKeySelective(approvalEventInfo);
